requirements.txt
ultralytics
streamlit
opencv-python-headless
numpy
pillow
